requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# AI Providers
openai>=1.12.0
anthropic>=0.18.0
ollama>=0.1.6

# PDF Processing
PyMuPDF>=1.23.0

# Git Operations
GitPython>=3.1.41

# Embeddings & Vector Store
sentence-transformers>=2.3.0
chromadb>=0.4.22

# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
httpx>=0.26.0
aiofiles>=23.2.1

# AST Parsing (Semantic Chunking)
tree-sitter>=0.23.0
tree-sitter-python>=0.23.0
tree-sitter-javascript>=0.23.0
tree-sitter-typescript>=0.23.0
tree-sitter-java>=0.23.0
tree-sitter-go>=0.23.0
tree-sitter-rust>=0.23.0
tree-sitter-c>=0.23.0
tree-sitter-cpp>=0.23.0
